requirements.txt
# Core AI and ML dependencies
torch>=2.0.0
torchvision
torchaudio
transformers>=4.30.0
accelerate
bitsandbytes

# Image processing
pillow>=9.0.0

# Vector database
chromadb>=0.4.0

# HTTP requests for Ollama
requests>=2.25.0

# Document processing
python-docx>=0.8.11

# Additional utilities
numpy
